+++ resources/views/pages/users/index.blade.php
@@ -1,5 +1,4 @@
 @use('Patrikjak\Starter\Policies\BasePolicy')
-@use('Patrikjak\Starter\Models\Users\Role')
 @use('Patrikjak\Starter\Models\Users\User')
 
 <x-pjstarter::layout.app :title="__('pjstarter::pages.users.title')">
@@ -10,13 +9,6 @@
                 @lang('pjstarter::pages.users.invite_user')
             </x-pjutils::button>
         @endcan
-
-        @can(BasePolicy::VIEW_ANY, Role::class)
-            <x-pjutils::button href="{{ route('admin.users.roles.index') }}" :bordered="true">
-                @lang('pjstarter::pages.users.roles.title')
-            </x-pjutils::button>
-        @endcan
-
 
     </x-slot:actions>