+++ src/Exceptions/Content/ContentImageStoreFailedException.php
@@ -10,6 +10,6 @@ final class ContentImageStoreFailedException extends RuntimeException
 {
     public function __construct()
     {
-        parent::__construct('Failed to store content image.');
+        parent::__construct('Failed to store content image');
     }
 }