+++ src/Services/Users/InviteService.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Collection;
 use Patrikjak\Auth\Services\InviteService as AuthInviteService;
 use Patrikjak\Starter\Models\Users\Role;
 use Patrikjak\Starter\Models\Users\User;
+use Patrikjak\Starter\Repositories\Contracts\Users\InvitationRepository;
 use Patrikjak\Starter\Repositories\Contracts\Users\RoleRepository;
 
 readonly class InviteService
@@ -17,12 +18,23 @@ readonly class InviteService
         private AuthInviteService $authInviteService,
         private RoleRepository $roleRepository,
         private AuthManager $authManager,
+        private InvitationRepository $invitationRepository,
     ) {
     }
 
     public function sendInvite(string $email, string $roleId): void
     {
         $this->authInviteService->sendInvite($email, $roleId);
+    }
+
+    public function delete(string $email): void
+    {
+        $this->invitationRepository->delete($email);
+    }
+
+    public function updateRole(string $email, string $roleId): void
+    {
+        $this->invitationRepository->updateRole($email, $roleId);
     }
 
     /**